# 03_openapi_original.yaml
openapi: 3.0.3
info:
  title: Kaveri Stays API
  version: 1.0.0
  description: API for Kaveri Stays
servers:
  - url: http://127.0.0.1:8000
...
paths:
...
  /bookings:
    get:
      tags: [Bookings]
      summary: List bookings
      parameters:
        - name: property_id
          in: query
          schema: {type: integer}
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/BookingStatus'
        - name: date_from
          in: query
          schema: {type: string, format: date}
        - name: date_to
          in: query
          schema: {type: string, format: date}
        - name: guest_id
          in: query
          schema: {type: integer}
        - name: sort
          in: query
          schema:
            type: string
            enum: [booking_id, check_in, check_out, created_at, status]
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: asc
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200': {description: Paginated bookings}
        '400': {description: Invalid filter or sort}
        '401': {description: Missing or invalid token}
        '403': {description: Access denied}

    post:
      tags: [Bookings]
      summary: Create a booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBookingRequest'
      responses:
        '201': {description: Booking created}
        '400': {description: Invalid request}
        '401': {description: Missing or invalid token}
        '403': {description: Not allowed}
        '409': {description: Room unavailable or constraint conflict}
...
  /bookings/{booking_id}/review:
    post:
      tags: [Reviews]
      summary: Leave a review
      parameters:
        - $ref: '#/components/parameters/BookingId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewRequest'
      responses:
        '201': {description: Review created}
        '400': {description: Invalid review}
        '401': {description: Missing or invalid token}
        '403': {description: Not the booking guest}
        '404': {description: Booking not found}
        '409': {description: Not checked out or review already exists}
...
components:
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    BookingId:
      name: booking_id
      in: path
      required: true
      schema: {type: integer, minimum: 1}
    Limit:
      name: limit
      in: query
      schema: {type: integer, minimum: 1, maximum: 100, default: 20}
    Offset:
      name: offset
      in: query
      schema: {type: integer, minimum: 0, default: 0}
...
  schemas:
...
    BookingStatus:
      type: string
      enum: [pending, confirmed, checked_in, checked_out, cancelled]
    CreateBookingRequest:
      type: object
      required: [property_id, room_id, check_in, check_out, guest_count, deposit_amount]
      properties:
        property_id: {type: integer}
        room_id: {type: integer}
        check_in: {type: string, format: date}
        check_out: {type: string, format: date}
        guest_count: {type: integer, minimum: 1}
        deposit_amount: {type: number, minimum: 0}
...
    ReviewRequest:
      type: object
      required: [rating]
      properties:
        rating: {type: integer, minimum: 1, maximum: 5}
        comment: {type: string, nullable: true}